+++ vital/kpf/CMakeLists.txt
@@ -4,7 +4,9 @@
 
 set( sources
   kpf_packet.cxx
+  kpf_parse_utils.cxx
   kpf_parse.cxx
+  kpf_text_parser.cxx
 )
 
 set( public_headers
@@ -12,14 +14,19 @@ set( public_headers
   kpf_canonical_types.h
   kpf_canonical_io.h
   kpf_parse.h
+  kpf_parse_utils.h
+  kpf_parser_base.h
+  kpf_text_parser.h
+  kpf_canonical_io_adapter_base.h
+  kpf_canonical_io_adapter.h
 )
 
 set( private_headers
-  kpf_parse_utils.h
+  kpf_bounce_buffer.h
 )
 
 set( private_sources
-  kpf_parse_utils.cxx
+  kpf_bounce_buffer.cxx
 )
 
 kwiver_add_library( vital_kpf
