+++ vital/kpf/examples/kpf_example_simple.cxx
@@ -9,6 +9,7 @@
 
 
 #include <vital/kpf/kpf_parse.h>
+#include <vital/kpf/kpf_text_parser.h>
 #include <iostream>
 #include <sstream>
 #include <string>
