+++ vital/kpf/kpf_parse.cxx
@@ -1,12 +1,10 @@
 #include "kpf_parse.h"
 
-#include <vital/util/tokenize.h>
 #include <vector>
 #include <stdexcept>
 
-#include <vital/kpf/kpf_parse_utils.h>
-
 #include <vital/logger/logger.h>
+#include <vital/kpf/kpf_canonical_io_adapter.h>
 
 using std::istream;
 using std::istringstream;
@@ -16,213 +14,16 @@ using std::pair;
 using std::make_pair;
 
 kwiver::vital::kpf::private_endl_t kwiver::vital::kpf::record_text_writer::endl;
-
-namespace { // anon
-using namespace kwiver::vital::kpf;
-
 static kwiver::vital::logger_handle_t main_logger( kwiver::vital::get_logger( __FILE__ ) );
-
-bool
-packet_header_parser( const string& s, packet_header_t& packet_header, bool expect_colon )
-{
-  //
-  // try to parse the header into a flag / tag / domain
-  //
-
-  header_parse_t h = parse_header( s, expect_colon );
-  if (! std::get<0>(h) )
-  {
-    return false;
-  }
-
-  string tag_str( std::get<1>(h) );
-  packet_style style = str2style( tag_str );
-  if ( style == packet_style::INVALID )
-  {
-    //    LOG_ERROR( main_logger, "Bad packet style '" << tag_str << "'" );
-    return false;
-  }
-
-  int domain( std::get<2>(h) );
-  packet_header = packet_header_t( style, domain );
-  return true;
-}
-
-bool
-packet_parser( const vector<string>& tokens,
-               packet_buffer_t& packet_buffer )
-{
-  size_t index(0), n( tokens.size() );
-
-  while ( index < n )
-  {
-    packet_t p;
-    if (packet_header_parser( tokens[ index ],
-                              p.header,
-                              true ))
-    {
-      // uh-oh, we couldn't parse it; build up an 'unparsed' key-value packet
-      // until we get a parse
-      ++index;
-      pair< bool, size_t > next = packet_payload_parser( index, tokens, p );
-      if (! next.first )
-      {
-        // This indicates a malformed packet error
-        return false;
-      }
-      index = next.second;
-
-      packet_buffer.insert( make_pair( p.header, p ));
-    }
-    else
-    {
-      // uh-oh, we couldn't recognize the header-- build up an 'unparsed' key-value
-      // packet
-      string unparsed_txt = tokens[index];
-      LOG_DEBUG( main_logger, "starting unparsed with '" << unparsed_txt << "'" );
-      // keep trying until we either parse a header or run out of tokens
-      ++index;
-      bool keep_going = (index < n);
-      while (keep_going)
-      {
-        if (packet_header_parser( tokens[index], p.header, true ))
-        {
-          // we found a parsable header-- all done
-          LOG_DEBUG( main_logger, "Found a parsable header at index " << index << ": '" << tokens[index] << "'" );
-          keep_going = false;
-        }
-        else
-        {
-          unparsed_txt += " "+tokens[index++];
-          keep_going = (index < n);
-        }
-      }
-
-      packet_header_t unparsed_header( packet_style::KV );
-      packet_t unparsed( unparsed_header );
-      LOG_DEBUG( main_logger, "Completing unparsed '" << unparsed_txt << "' ; next index " << index << " of " << n);
-      new (&unparsed.kv) canonical::kv_t( "unparsed", unparsed_txt );
-      packet_buffer.insert( make_pair( unparsed.header, unparsed ));
-    }
-  }
-  return true;
-}
-
-
-} // ...anon
 
 namespace kwiver {
 namespace vital {
 namespace kpf {
 
-//
-// text parser
-//
-
-kpf_text_parser_t
-::kpf_text_parser_t( istream& is ): input_stream( is )
-{
-}
-
-bool
-kpf_text_parser_t
-::get_status() const
-{
-  return static_cast<bool>( this->input_stream );
-}
-
-bool
-kpf_text_parser_t
-::parse_next_record( packet_buffer_t& local_packet_buffer )
-{
-  string s;
-  if (! std::getline( this->input_stream, s ))
-  {
-    return false;
-  }
-  vector< string > tokens;
-  ::kwiver::vital::tokenize( s, tokens, " ", true );
-
-  bool rc = packet_parser( tokens, local_packet_buffer );
-  return rc;
-}
-
-//
-//
-//
-
-packet_bounce_t
-::packet_bounce_t()
-  : is_set( false )
-{
-}
-
-packet_bounce_t
-::packet_bounce_t( const string& s )
-  : is_set( false )
-{
-  this->init( s );
-}
-
-packet_bounce_t
-::packet_bounce_t( const packet_header_t& h )
-  : is_set( false ), header( h )
-{
-}
-
-void
-packet_bounce_t
-::init( const string& s )
-{
-  if (! packet_header_parser( s, this->header, false ))
-  {
-    LOG_ERROR( main_logger, "Couldn't create a reader for packets of type '" << s << "'");
-    return;
-  }
-}
-
-void packet_bounce_t
-::init( const packet_header_t& h )
-{
-  this->header = h;
-}
-
-void
-packet_bounce_t
-::set_from_buffer( const packet_t& p )
-{
-  this->is_set = true;
-  this->packet = p;
-}
-
-packet_header_t
-packet_bounce_t
-::my_header() const
-{
-  return this->header;
-}
-
-packet_bounce_t&
-packet_bounce_t
-::set_domain( int d )
-{
-  this->is_set = false;
-  this->header.domain = d;
-  return *this;
-}
-
-pair< bool, packet_t >
-packet_bounce_t
-::get_packet()
-{
-  bool f = this->is_set;
-  this->is_set = false;
-  return make_pair( f, this->packet );
-}
 
 kpf_reader_t
 ::kpf_reader_t( kpf_parser_base_t& p )
-  : packet_buffer( packet_header_cmp ), parser(p)
+  : parser(p)
 {
   this->reader_status = this->parser.get_status();
 }
@@ -267,7 +68,7 @@ kpf_reader_t
   packet_header_t meta_packet_h( packet_style::META );
   while ( ! non_meta_packets_added )
   {
-    packet_buffer_t local_packet_buffer( packet_header_cmp );
+    packet_buffer_t local_packet_buffer;
     bool rc = this->parser.parse_next_record( local_packet_buffer );
     if (! rc )
     {
@@ -427,15 +228,15 @@ operator>>( kpf_reader_t& t,
 
 bool
 kpf_reader_t
-::process( kpf_io_adapter_base& io )
+::process( kpf_canonical_io_adapter_base& io )
 {
-  return this->process( io.text_reader );
+  return this->process( io.packet_bounce );
 }
 
 kpf_reader_t& operator>>( kpf_reader_t& t,
-                           kpf_io_adapter_base& io )
+                          kpf_canonical_io_adapter_base& io )
 {
-  return t >> io.text_reader;
+  return t >> io.packet_bounce;
 }
 
 kpf_reader_t&
