+++ vital/kpf/kpf_parse_utils.cxx
@@ -193,7 +193,6 @@ namespace kwiver {
 namespace vital {
 namespace kpf {
 
-
 //
 // Given a string which is expected to be a packet header (e.g.
 // 'g0:', 'meta:', 'eval19:') separate it into a success flag,
@@ -332,7 +331,92 @@ packet_payload_parser ( size_t index,
 
 }
 
+bool
+packet_header_parser( const string& s,
+                      packet_header_t& packet_header,
+                      bool expect_colon )
+{
+  //
+  // try to parse the header into a flag / tag / domain
+  //
 
+  header_parse_t h = parse_header( s, expect_colon );
+  if (! std::get<0>(h) )
+  {
+    return false;
+  }
+
+  string tag_str( std::get<1>(h) );
+  packet_style style = str2style( tag_str );
+  if ( style == packet_style::INVALID )
+  {
+    //    LOG_ERROR( main_logger, "Bad packet style '" << tag_str << "'" );
+    return false;
+  }
+
+  int domain( std::get<2>(h) );
+  packet_header = packet_header_t( style, domain );
+  return true;
+}
+
+bool
+packet_parser( const vector<string>& tokens,
+               packet_buffer_t& packet_buffer )
+{
+  size_t index(0), n( tokens.size() );
+  while ( index < n )
+  {
+    packet_t p;
+    if (packet_header_parser( tokens[ index ],
+                              p.header,
+                              true ))
+    {
+      // uh-oh, we couldn't parse it; build up an 'unparsed' key-value packet
+      // until we get a parse
+      ++index;
+      pair< bool, size_t > next = packet_payload_parser( index, tokens, p );
+      if (! next.first )
+      {
+        // This indicates a malformed packet error
+        return false;
+      }
+      index = next.second;
+
+      packet_buffer.insert( make_pair( p.header, p ));
+    }
+    else
+    {
+      // uh-oh, we couldn't recognize the header-- build up an 'unparsed' key-value
+      // packet
+      string unparsed_txt = tokens[index];
+      LOG_DEBUG( main_logger, "starting unparsed with '" << unparsed_txt << "'" );
+      // keep trying until we either parse a header or run out of tokens
+      ++index;
+      bool keep_going = (index < n);
+      while (keep_going)
+      {
+        if (packet_header_parser( tokens[index], p.header, true ))
+        {
+          // we found a parsable header-- all done
+          LOG_DEBUG( main_logger, "Found a parsable header at index " << index << ": '" << tokens[index] << "'" );
+          keep_going = false;
+        }
+        else
+        {
+          unparsed_txt += " "+tokens[index++];
+          keep_going = (index < n);
+        }
+      }
+
+      packet_header_t unparsed_header( packet_style::KV );
+      packet_t unparsed( unparsed_header );
+      LOG_DEBUG( main_logger, "Completing unparsed '" << unparsed_txt << "' ; next index " << index << " of " << n);
+      new (&unparsed.kv) canonical::kv_t( "unparsed", unparsed_txt );
+      packet_buffer.insert( make_pair( unparsed.header, unparsed ));
+    }
+  }
+  return true;
+}
 
 } // ...kpf
 } // ...vital
